updates from local copy

+++ _scripts/setup_java_ssl.bash
@@ -13,6 +13,12 @@ echo "Using temp directory $tmp_dir"
 
 curl -o "$tmp_dir/doi-cacert.cer" https://apps-int.usgs.gov/ssl/DOIRootCA2.cer
 
+if [ $? -ne 0 ]; then
+    echo "Unable to fetch the cert file.  curl error code $?"
+    exit 1
+fi
+
+
 echo "Attempting to delete the cert aliased as 'DOI_Cert', which may error if it doesn't exist"
 keytool -delete -alias DOI_Cert -cacerts -storepass changeit
 
@@ -20,6 +26,10 @@ echo "Attempting to add new certificate, which may cause a warning if it already
 echo "If it does exist, remove via > keytool -delete -alias [TheAlias] -cacerts -storepass changeit"
 keytool -importcert -file "$tmp_dir/doi-cacert.cer" -alias DOI_Cert -cacerts -storepass changeit -noprompt
 
+if [ $? -ne 0 ]; then
+    echo "Unable to add - If a permissions issues, this should be run with sudo.  Error code $?"
+    exit 1
+fi
 
 echo "Removing temp directory..."
 rm -r $tmp_dir
